let's get logout to hit an endpoint
