+++ gcf/admin/users/main.py
@@ -1,8 +1,7 @@
 from flask import Request, Response, jsonify
 from google.cloud import firestore_v1
-from firebase_admin import firestore
 
-db = firestore.Client()
+db = firestore_v1.Client(database="grindolympiads")
 
 
 def list_users(request: Request) -> Response:
