+++ gcf/admin/users/requirements.txt
@@ -1,4 +1,3 @@
-firebase-admin
 Flask>=1.0,<3.0
 functions-framework==3.0.0
 google-cloud-firestore==2.9.1
