+++ gcf/login/requirements.txt
@@ -1,8 +1,8 @@
-google-auth
-google-auth-oauthlib
-google-auth-httplib2
-google-cloud-firestore==2.9.1
 firebase-admin
-functions-framework==3.0.0
 Flask>=1.0,<3.0
+functions-framework==3.0.0
+google-auth
+google-cloud-firestore==2.9.1
+google-auth-httplib2
+google-auth-oauthlib
 pyjwt==2.8.0
